# sdk/typescript/openapi.json
{
  "components": {
    "schemas": {
      "CompilationCandidate": {
        "additionalProperties": false,
        "description": "Declarative candidate; it is safe to evaluate, not auto-promote.",
        "properties": {
          "candidate_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Candidate Digest",
            "type": "string"
          },
          "dependencies": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DependencyEvidence"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "effect_scope": {
            "const": "pure_or_read",
            "default": "pure_or_read",
            "title": "Effect Scope",
            "type": "string"
          },
          "intent": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Intent",
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/CompiledNode"
            },
            "minItems": 1,
            "title": "Nodes",
            "type": "array"
          },
          "schema_version": {
            "const": "awe.candidate.v1",
            "default": "awe.candidate.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_trace_ids": {
            "items": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
              "type": "string"
            },
            "minItems": 2,
            "title": "Source Trace Ids",
            "type": "array"
          }
        },
        "required": [
          "candidate_digest",
          "intent",
          "source_trace_ids",
          "nodes"
        ],
        "title": "CompilationCandidate",
        "type": "object"
      },
      "CompilationReceipt": {
        "additionalProperties": false,
        "description": "Deterministic receipt for either compilation or a safe refusal.",
        "properties": {
          "candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompilationCandidate"
              },
              {
                "type": "null"
              }
            ]
          },
          "compiler_version": {
            "const": "awe.compiler.v1",
            "default": "awe.compiler.v1",
            "title": "Compiler Version",
            "type": "string"
          },
          "input_bundle_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Input Bundle Digest",
            "type": "string"
          },
          "reasons": {
            "default": [],
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Receipt Hash",
            "type": "string"
          },
          "schema_version": {
            "const": "awe.compilation-receipt.v1",
            "default": "awe.compilation-receipt.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "compiled",
              "refused"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "input_bundle_digest",
          "status",
          "receipt_hash"
        ],
        "title": "CompilationReceipt",
        "type": "object"
      },
      "CompileRequest": {
        "additionalProperties": false,
        "description": "Trace set proposed for compilation.",
        "properties": {
          "traces": {
            "items": {
              "$ref": "#/components/schemas/ExecutionTrace"
            },
            "maxItems": 128,
            "minItems": 1,
            "title": "Traces",
            "type": "array"
          }
        },
        "required": [
          "traces"
        ],
        "title": "CompileRequest",
        "type": "object"
      },
...
      "EvaluateRequest": {
        "additionalProperties": false,
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/EvaluationBundle"
          },
          "candidate": {
            "$ref": "#/components/schemas/EvaluationBundle"
          },
          "policy": {
            "$ref": "#/components/schemas/EvaluationPolicy",
            "default": {
              "maximum_cost_increase_bps": 2500,
              "maximum_latency_increase_bps": 2500,
              "maximum_success_regression_bps": 200,
              "minimum_trials": 3,
              "require_zero_safety_violations": true,
              "schema_version": "awe.evaluation-policy.v1"
            }
          }
        },
        "required": [
          "baseline",
          "candidate"
        ],
        "title": "EvaluateRequest",
        "type": "object"
      },
...
      "EvaluationReceipt": {
        "additionalProperties": false,
        "description": "Fail-closed comparison of a baseline and candidate on one frozen dataset.",
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/EvaluationMetrics"
          },
          "baseline_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Baseline Digest",
            "type": "string"
          },
          "candidate": {
            "$ref": "#/components/schemas/EvaluationMetrics"
          },
          "candidate_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Candidate Digest",
            "type": "string"
          },
          "dataset_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Dataset Digest",
            "type": "string"
          },
          "evaluator_version": {
            "const": "awe.evaluator.v1",
            "default": "awe.evaluator.v1",
            "title": "Evaluator Version",
            "type": "string"
          },
          "policy_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Policy Digest",
            "type": "string"
          },
          "reasons": {
            "default": [],
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Receipt Hash",
            "type": "string"
          },
          "schema_version": {
            "const": "awe.evaluation-receipt.v1",
            "default": "awe.evaluation-receipt.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "pass",
              "review",
              "block"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "baseline_digest",
          "candidate_digest",
          "dataset_digest",
          "policy_digest",
          "status",
          "baseline",
          "candidate",
          "receipt_hash"
        ],
        "title": "EvaluationReceipt",
        "type": "object"
      },
...
      "ExecutionTrace": {
        "additionalProperties": false,
        "description": "Immutable evidence from a single workflow execution.",
        "properties": {
          "intent": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_]*$",
            "title": "Intent",
            "type": "string"
          },
          "schema_version": {
            "const": "awe.trace.v1",
            "default": "awe.trace.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TraceStep"
            },
            "maxItems": 256,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          },
          "succeeded": {
            "title": "Succeeded",
            "type": "boolean"
          },
          "trace_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Trace Id",
            "type": "string"
          },
          "workflow_inputs": {
            "items": {
              "$ref": "#/components/schemas/FieldEvidence"
            },
            "maxItems": 256,
            "minItems": 1,
            "title": "Workflow Inputs",
            "type": "array"
          }
        },
        "required": [
          "trace_id",
          "intent",
          "succeeded",
          "workflow_inputs",
          "steps"
        ],
        "title": "ExecutionTrace",
        "type": "object"
      },
      "ExperimentManifest": {
        "additionalProperties": false,
        "description": "Content-addressed experiment evidence imported from one pinned format.",
        "properties": {
          "commit_sha": {
            "pattern": "^(?:[0-9a-f]{40}|[0-9a-f]{64})$",
            "title": "Commit Sha",
            "type": "string"
          },
          "dataset_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Dataset Digest",
            "type": "string"
          },
          "dataset_split_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Dataset Split Digest",
            "type": "string"
          },
          "environment_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Environment Digest",
            "type": "string"
          },
          "experiment_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Experiment Id",
            "type": "string"
          },
          "grader_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Grader Digest",
            "type": "string"
          },
          "harness_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Harness Digest",
            "type": "string"
          },
          "harness_name": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Harness Name",
            "type": "string"
          },
          "harness_version": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9.+_-]*$",
            "title": "Harness Version",
            "type": "string"
          },
          "manifest_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Manifest Digest",
            "type": "string"
          },
          "model_config_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Model Config Digest",
            "type": "string"
          },
          "model_name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "model_provider": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Model Provider",
            "type": "string"
          },
          "repository_uri": {
            "maxLength": 512,
            "minLength": 12,
            "pattern": "^https://[A-Za-z0-9.-]+(?::[0-9]{1,5})?/[A-Za-z0-9._~!$&'()*+,;=:@%/-]+$",
            "title": "Repository Uri",
            "type": "string"
          },
          "schema_version": {
            "const": "awe.experiment-manifest.v1",
            "default": "awe.experiment-manifest.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_format": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Source Format",
            "type": "string"
          },
          "source_revision": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9.+_-]*$",
            "title": "Source Revision",
            "type": "string"
          },
          "strategy_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Strategy Digest",
            "type": "string"
          },
          "strategy_name": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Strategy Name",
            "type": "string"
          },
          "subject_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Subject Digest",
            "type": "string"
          },
          "trials": {
            "items": {
              "$ref": "#/components/schemas/ExperimentTrial"
            },
            "maxItems": 100000,
            "minItems": 1,
            "title": "Trials",
            "type": "array"
          }
        },
        "required": [
          "experiment_id",
          "repository_uri",
          "commit_sha",
          "subject_digest",
          "dataset_digest",
          "dataset_split_digest",
          "harness_name",
          "harness_version",
          "harness_digest",
          "strategy_name",
          "strategy_digest",
          "model_provider",
          "model_name",
          "model_config_digest",
          "environment_digest",
          "grader_digest",
          "trials",
          "source_format",
          "source_revision",
          "manifest_digest"
        ],
        "title": "ExperimentManifest",
        "type": "object"
      },
      "ExperimentRun": {
        "additionalProperties": false,
        "description": "Provider-neutral evidence shared by generic and telemetry importers.",
        "properties": {
          "commit_sha": {
            "pattern": "^(?:[0-9a-f]{40}|[0-9a-f]{64})$",
            "title": "Commit Sha",
            "type": "string"
          },
          "dataset_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Dataset Digest",
            "type": "string"
          },
          "dataset_split_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Dataset Split Digest",
            "type": "string"
          },
          "environment_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Environment Digest",
            "type": "string"
          },
          "experiment_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Experiment Id",
            "type": "string"
          },
          "grader_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Grader Digest",
            "type": "string"
          },
          "harness_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Harness Digest",
            "type": "string"
          },
          "harness_name": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Harness Name",
            "type": "string"
          },
          "harness_version": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9.+_-]*$",
            "title": "Harness Version",
            "type": "string"
          },
          "model_config_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Model Config Digest",
            "type": "string"
          },
          "model_name": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "model_provider": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Model Provider",
            "type": "string"
          },
          "repository_uri": {
            "maxLength": 512,
            "minLength": 12,
            "pattern": "^https://[A-Za-z0-9.-]+(?::[0-9]{1,5})?/[A-Za-z0-9._~!$&'()*+,;=:@%/-]+$",
            "title": "Repository Uri",
            "type": "string"
          },
          "strategy_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Strategy Digest",
            "type": "string"
          },
          "strategy_name": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "title": "Strategy Name",
            "type": "string"
          },
          "subject_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Subject Digest",
            "type": "string"
          },
          "trials": {
            "items": {
              "$ref": "#/components/schemas/ExperimentTrial"
            },
            "maxItems": 100000,
            "minItems": 1,
            "title": "Trials",
            "type": "array"
          }
        },
        "required": [
          "experiment_id",
          "repository_uri",
          "commit_sha",
          "subject_digest",
          "dataset_digest",
          "dataset_split_digest",
          "harness_name",
          "harness_version",
          "harness_digest",
          "strategy_name",
          "strategy_digest",
          "model_provider",
          "model_name",
          "model_config_digest",
          "environment_digest",
          "grader_digest",
          "trials"
        ],
        "title": "ExperimentRun",
        "type": "object"
      },
...
      "FieldEvidence": {
        "additionalProperties": false,
        "description": "Content digest observed at a JSON field boundary.",
        "properties": {
          "field": {
            "maxLength": 256,
            "minLength": 2,
            "pattern": "^/(?:[^~\\x00-\\x1f]|~[01])+$",
            "title": "Field",
            "type": "string"
          },
          "value_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Value Digest",
            "type": "string"
          }
        },
        "required": [
          "field",
          "value_digest"
        ],
        "title": "FieldEvidence",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromotionReceipt": {
        "additionalProperties": false,
        "description": "Human decision bound to a verified compilation and evaluation chain.",
        "properties": {
          "actor_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:@+-]*$",
            "title": "Actor Id",
            "type": "string"
          },
          "candidate_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Candidate Digest",
            "type": "string"
          },
          "commit_sha": {
            "pattern": "^(?:[0-9a-f]{40}|[0-9a-f]{64})$",
            "title": "Commit Sha",
            "type": "string"
          },
          "compilation_receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Compilation Receipt Hash",
            "type": "string"
          },
          "dataset_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Dataset Digest",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evaluation_receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Evaluation Receipt Hash",
            "type": "string"
          },
          "evaluation_status": {
            "enum": [
              "pass",
              "review",
              "block"
            ],
            "title": "Evaluation Status",
            "type": "string"
          },
          "input_bundle_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Input Bundle Digest",
            "type": "string"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "policy_digest": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Policy Digest",
            "type": "string"
          },
          "rationale": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Receipt Hash",
            "type": "string"
          },
          "schema_version": {
            "const": "awe.promotion-receipt.v2",
            "default": "awe.promotion-receipt.v2",
            "title": "Schema Version",
            "type": "string"
          },
          "traces_verified": {
            "title": "Traces Verified",
            "type": "boolean"
          },
          "verification_receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Verification Receipt Hash",
            "type": "string"
          },
          "verification_status": {
            "enum": [
              "valid",
              "invalid"
            ],
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "candidate_digest",
          "compilation_receipt_hash",
          "input_bundle_digest",
          "verification_receipt_hash",
          "verification_status",
          "traces_verified",
          "evaluation_receipt_hash",
          "evaluation_status",
          "dataset_digest",
          "policy_digest",
          "decision",
          "actor_id",
          "commit_sha",
          "issued_at",
          "rationale",
          "receipt_hash"
        ],
        "title": "PromotionReceipt",
        "type": "object"
      },
      "PromotionRequest": {
        "additionalProperties": false,
        "description": "Evidence inputs required to record a human promotion decision.",
        "properties": {
          "actor_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:@+-]*$",
            "title": "Actor Id",
            "type": "string"
          },
          "commit_sha": {
            "pattern": "^(?:[0-9a-f]{40}|[0-9a-f]{64})$",
            "title": "Commit Sha",
            "type": "string"
          },
          "compilation": {
            "$ref": "#/components/schemas/CompilationReceipt"
          },
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ],
            "title": "Decision",
            "type": "string"
          },
          "evaluation": {
            "$ref": "#/components/schemas/EvaluationReceipt"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "rationale": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "traces": {
            "items": {
              "$ref": "#/components/schemas/ExecutionTrace"
            },
            "title": "Traces",
            "type": "array"
          },
          "verification": {
            "$ref": "#/components/schemas/ReceiptVerification"
          }
        },
        "required": [
          "compilation",
          "verification",
          "traces",
          "evaluation",
          "decision",
          "actor_id",
          "commit_sha",
          "issued_at",
          "rationale"
        ],
        "title": "PromotionRequest",
        "type": "object"
      },
      "ReceiptVerification": {
        "additionalProperties": false,
        "description": "Result of recomputing a compilation receipt and optional source traces.",
        "properties": {
          "reasons": {
            "default": [],
            "items": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "receipt_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Receipt Hash",
            "type": "string"
          },
          "schema_version": {
            "const": "awe.receipt-verification.v2",
            "default": "awe.receipt-verification.v2",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "valid",
              "invalid"
            ],
            "title": "Status",
            "type": "string"
          },
          "traces_verified": {
            "title": "Traces Verified",
            "type": "boolean"
          },
          "verification_hash": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Verification Hash",
            "type": "string"
          }
        },
        "required": [
          "status",
          "receipt_hash",
          "traces_verified",
          "verification_hash"
        ],
        "title": "ReceiptVerification",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "receipt": {
            "$ref": "#/components/schemas/CompilationReceipt"
          },
          "traces": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ExecutionTrace"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traces"
          }
        },
        "required": [
          "receipt"
        ],
        "title": "VerifyRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Evidence-gated, read-only trace compiler and verifier",
    "title": "AWE TraceGate",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/compile": {
      "post": {
        "operationId": "compile_workflow_v1_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile Workflow"
      }
    },
    "/v1/evaluate": {
      "post": {
        "operationId": "evaluate_v1_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate"
      }
    },
    "/v1/experiments/import/generic": {
      "post": {
        "operationId": "import_experiment_v1_experiments_import_generic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRun"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Experiment"
      }
    },
    "/v1/promote": {
      "post": {
        "operationId": "promote_v1_promote_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromotionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromotionReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote"
      }
    },
    "/v1/verify": {
      "post": {
        "operationId": "verify_receipt_v1_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptVerification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Receipt"
      }
    }
  }
}